Hopefully this resolves the merge conflicts, partially refactored peripheralListActivity
