+++ Hestia/app/src/main/java/com/rugged/application/hestia/DeviceListActivity.java
@@ -10,10 +10,10 @@ import android.support.v4.app.Fragment;
  * The activity which presents a list containing all peripherals to the user. An activity is a
  * single, focused thing the user can do.
  */
-public class PeripheralListActivity extends SingleFragmentActivity{
-    private final static String TAG = "PeripheralListActivity";
+public class DeviceListActivity extends SingleFragmentActivity{
+    private final static String TAG = "DeviceListActivity";
     @Override
     protected Fragment createFragment() {
-        return new PeripheralListFragment();
+        return new DeviceListFragment();
     }
 }
