--- Hestia/app/src/main/java/com/rugged/application/hestia/PeripheralListFragment.java
@@ -1,231 +1,0 @@
-package com.rugged.application.hestia;
-import android.content.Context;
-import android.graphics.Typeface;
-import android.os.AsyncTask;
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.PopupMenu;
-import android.view.LayoutInflater;
-import android.view.MenuItem;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.BaseExpandableListAdapter;
-import android.widget.ExpandableListView;
-import android.widget.ImageView;
-import android.widget.TextView;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-
-/**
- * This fragment takes care of generating the list of peripherals on the phone. It sends an HTTP
- * GET request to the server to populate the device list.
- *
- * @see PeripheralListActivity
- */
-public class PeripheralListFragment extends Fragment {
-
-    private final static String TAG = "PeripheralListFragment";
-
-    ExpandableListAdapter listAdapter;
-    ExpandableListView expListView;
-    List<String> listDataHeader;
-    HashMap<String, List<Device>> listDataChild;
-
-    /**
-     *
-     * @param inflater The layout inflater used to generate the layout hierarchy
-     * @param container The viewgroup with which the layout is instantiated
-     * @return A view of an expandable list linked to the listDataHeader and listDataChild
-     *         variables. Filling these lists will generate the GUI.
-     */
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_peripheral_list, container, false);
-
-        expListView = (ExpandableListView) view.findViewById(R.id.lvExp);
-
-        listDataHeader = new ArrayList<>();
-        listDataChild = new HashMap<>();
-
-        listAdapter = new ExpandableListAdapter(listDataHeader, listDataChild);
-
-        expListView.setAdapter(listAdapter);
-
-        //request the list
-        new RetrievePeripheralList().execute();
-
-        // setting list adapter
-        return view;
-    }
-
-    private class ExpandableListAdapter extends BaseExpandableListAdapter {
-        private List<String> listDataHeader; // header titles
-        // child data in format of header title, child title
-        private HashMap<String, List<Device>> listDataChild;
-
-        private ExpandableListAdapter(List<String> listDataHeader,
-                                     HashMap<String, List<Device>> listChildData) {
-            this.listDataHeader = listDataHeader;
-            this.listDataChild = listChildData;
-        }
-
-        @Override
-        public Object getChild(int groupPosition, int childPosititon) {
-            return this.listDataChild.get(this.listDataHeader.get(groupPosition))
-                    .get(childPosititon).getName();
-        }
-
-        public Device getChildDevice(int groupPostition, int childPostion) {
-            return this.listDataChild.get(this.listDataHeader.get(groupPostition))
-                    .get(childPostion);
-        }
-
-        @Override
-        public long getChildId(int groupPosition, int childPosition) {
-            return childPosition;
-        }
-
-        @Override
-        public View getChildView(final int groupPosition, final int childPosition,
-                                 boolean isLastChild, View convertView, ViewGroup parent) {
-            final String childText = (String) getChild(groupPosition, childPosition);
-
-            if (convertView == null) {
-                LayoutInflater infalInflater = (LayoutInflater) getActivity()
-                        .getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-                //inflate depending on toggle or slider
-                convertView = infalInflater.inflate(R.layout.child_list_item, null);
-            }
-
-            TextView txtListChild = (TextView) convertView.findViewById(R.id.child_item_text);
-            txtListChild.setText(childText);
-
-            ImageView imageview = (ImageView) convertView.findViewById(R.id.imageview);
-
-            imageview.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    PopupMenu popup = new PopupMenu(getActivity(), view);
-                    popup.getMenuInflater().inflate(R.menu.popup,
-                            popup.getMenu());
-
-                    for (int i = 0; i < getChildDevice(groupPosition, childPosition)
-                            .getActivators().size(); i++) {
-                        if (getChildDevice(groupPosition, childPosition).getActivators().get(i)
-                                .getType().equals("SLIDER")) {
-                            popup.getMenu().findItem(R.id.slide).setEnabled(true);
-                            popup.getMenu().findItem(R.id.slide).setVisible(true);
-                            break;
-                        }
-                    }
-
-                    popup.show();
-                    popup.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
-                        @Override
-                        public boolean onMenuItemClick(MenuItem item) {
-                            switch (item.getItemId()) {
-                                case R.id.settings :
-                                    //Settings
-                                    break;
-                                case R.id.delete :
-                                    //Remove list item
-                                    break;
-                                case R.id.slide :
-                                    //show notification
-                                    final SlideDialog dialog = new SlideDialog(getActivity());
-                                    dialog.show();
-                                    break;
-                                default:
-                                    break;
-                            }
-                            return true;
-                        }
-                    });
-                }
-            });
-            return convertView;
-        }
-
-        @Override
-        public int getChildrenCount(int groupPosition) {
-            return this.listDataChild.get(this.listDataHeader.get(groupPosition))
-                    .size();
-        }
-
-        @Override
-        public Object getGroup(int groupPosition) {
-            return this.listDataHeader.get(groupPosition);
-        }
-
-        @Override
-        public int getGroupCount() {
-            return this.listDataHeader.size();
-        }
-
-        @Override
-        public long getGroupId(int groupPosition) {
-            return groupPosition;
-        }
-
-        @Override
-        public View getGroupView(int groupPosition, boolean isExpanded,
-                                 View convertView, ViewGroup parent) {
-            String headerTitle = (String) getGroup(groupPosition);
-            if (convertView == null) {
-                LayoutInflater infalInflater = (LayoutInflater) getActivity()
-                        .getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-
-                convertView = infalInflater.inflate(R.layout.list_group, null);
-            }
-
-            TextView lblListHeader = (TextView) convertView
-                    .findViewById(R.id.lblListHeader);
-            lblListHeader.setTypeface(null, Typeface.BOLD);
-            lblListHeader.setText(headerTitle);
-
-            return convertView;
-        }
-
-        @Override
-        public boolean hasStableIds() {
-            return false;
-        }
-
-        @Override
-        public boolean isChildSelectable(int groupPosition, int childPosition) {
-            return true;
-        }
-    }
-
-    private class RetrievePeripheralList extends AsyncTask<Void, Void, Void>  {
-        @Override
-        protected Void doInBackground(Void... voids)  {
-            //mock data
-            ArrayList<Activator> activators = new ArrayList<>();
-            activators.add(new Activator<>(0, false, "light_OnOROff", "TOGGLE"));
-            ArrayList<Activator> a2 = new ArrayList<>();
-            a2.add(new Activator<>(0, 0, "Lock_OnOROff", "SLIDER"));
-            Device d1 = new Device(0, "Light 1", "Light", activators);
-            Device d2 = new Device(0, "Light 2", "Light", activators);
-            Device d3 = new Device(0, "lock 1", "Lock", a2);
-            ArrayList<Device> devices = new ArrayList<>();
-            devices.add(d1);
-            devices.add(d2);
-            devices.add(d3);
-
-            // add header data
-            for (Device d : devices) {
-                if (!listDataHeader.contains(d.getType())) {
-                    listDataHeader.add(d.getType());
-                    listDataChild.put(d.getType(), new ArrayList<Device>());
-                }
-                //find corresponding header for the child
-                listDataChild.get(d.getType()).add(d);
-            }
-//            listAdapter.notifyDataSetChanged();
-            return null;
-        }
-    }
-}
